Updated to reflect new standard for Child Themeing
According to http://codex.wordpress.org/Child_Themes#How_to_Create_a_Child_Theme The new rule is to not @import the css file directly but use functions.php to enqueue the old css file. This reflects that and should fix broken-ness in ChildThemeing new themes in the current version of WordPress.