Updated to reflect new standard for Child Themeing
According to http://codex.wordpress.org/Child_Themes#How_to_Create_a_Child_Theme The new rule is to not @import the css file directly but use functions.php to enqueue the old css file. This reflects that and should fix broken-ness in ChildThemeing new themes in the current version of WordPress.

git-svn-id: http://plugins.svn.wordpress.org/one-click-child-theme/trunk@1117387 b8457f37-d9ea-0310-8a92-e5e31aec5664